\name{prediction_values}
\alias{prediction_values}
\docType{data}
\title{
Test data produced by \code{predict}
}
\description{
This is the output from running the \code{predict} function on 
\code{testdata}. It has been added to make the \code{summary}
example legible. 
}
\usage{data(prediction_values)}

\examples{
data(prediction_values)
}
\keyword{datasets}
