\name{testdata}
\alias{testdata}
\docType{data}
\title{
Validation data set for the VOrbitrap machine  
}
\description{
This is a validation training set for the LLRC composed 
of 100 unclassified proteomics samples from the Vorbitrap machine. 
}
\usage{data(testdata)}
\examples{data(testdata)}
\keyword{datasets}
